+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+### New in 0.1.6 (Released 2016/02/29)
+* Add prefix to GetObjects to limit results
+* Rename GetObject to "GetS3Object"
+
 ### New in 0.1.6 (Released 2016/02/27)
 * GetBytes